generalize sliderView options 1 and 2
